% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{normN}
\alias{normN}
\title{Calculate the norm value of each element in the list data}
\usage{
normN(data)
}
\arguments{
\item{data}{The list data.}
}
\description{
Calculate the norm value of each element in the list data
}
