% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ParallelCoordinatesPlot}
\alias{ParallelCoordinatesPlot}
\title{plots a simple Parallel Coordinates Plot}
\usage{
ParallelCoordinatesPlot(data, col = 2, lwd = 1, lty = 1, ...)
}
\arguments{
\item{data}{The list data for plot.}

\item{col}{The count of columns.}

\item{lwd}{The line width.}

\item{lty}{The line type.}

\item{...}{a list of tagged values.}
}
\description{
plots a simple Parallel Coordinates Plot
}
\examples{
data<-list(setosa=iris[iris$Species=="setosa","Sepal.Length"],
versicolor=iris[iris$Species=="versicolor","Sepal.Length"],
virginica=iris[iris$Species=="virginica","Sepal.Length"])
ParallelCoordinatesPlot(data,col=3)
}
